# schema/openapi/openapi.yaml
openapi: '3.0.3'
info:
  title: Backbone API
  version: '1.0.0'
  description: |
    Auto-generated by backbone-cli.
    This API provides 11 standard CRUD endpoints for each model.
servers:
  - url: http://localhost:3000
    description: Development server
...
paths:
...
  /api/v1/pos_profiles:
    get:
      tags:
        - PosProfiles
      summary: List PosProfiles
      description: Get a paginated list of PosProfiles with optional filtering and sorting
      operationId: listPosProfiles
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PosProfileList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - PosProfiles
      summary: Create PosProfile
      description: Create a new PosProfile
      operationId: createPosProfile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePosProfileRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PosProfile'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  schemas:
...
    PosProfile:
      type: object
      required:
        - id
        - company_id
        - name
        - currency
        - tax_rate
        - allow_discount
        - is_active
        - metadata
      properties:
        id:
          type: string
          format: uuid
        company_id:
          type: string
          format: uuid
        branch_id:
          nullable: true
          type: string
          format: uuid
        name:
          type: string
          maxLength: 140
        default_customer_id:
          nullable: true
          type: string
          format: uuid
        currency:
          type: string
          maxLength: 3
        income_account_id:
          nullable: true
          type: string
          format: uuid
        receivable_account_id:
          nullable: true
          type: string
          format: uuid
        cash_account_id:
          nullable: true
          type: string
          format: uuid
        write_off_account_id:
          nullable: true
          type: string
          format: uuid
        tax_account_id:
          nullable: true
          type: string
          format: uuid
        tax_rate:
          type: string
        warehouse_id:
          nullable: true
          type: string
          format: uuid
        cogs_account_id:
          nullable: true
          type: string
          format: uuid
        inventory_account_id:
          nullable: true
          type: string
          format: uuid
        allow_discount:
          type: boolean
        is_active:
          type: boolean
        metadata:
          type: object
...
    CreatePosProfileRequest:
      type: object
      required:
        - company_id
        - name
        - currency
        - tax_rate
        - allow_discount
        - is_active
        - metadata
      properties:
        company_id:
          type: string
          format: uuid
        branch_id:
          nullable: true
          type: string
          format: uuid
        name:
          type: string
          maxLength: 140
        default_customer_id:
          nullable: true
          type: string
          format: uuid
        currency:
          type: string
          maxLength: 3
        income_account_id:
          nullable: true
          type: string
          format: uuid
        receivable_account_id:
          nullable: true
          type: string
          format: uuid
        cash_account_id:
          nullable: true
          type: string
          format: uuid
        write_off_account_id:
          nullable: true
          type: string
          format: uuid
        tax_account_id:
          nullable: true
          type: string
          format: uuid
        tax_rate:
          type: string
        warehouse_id:
          nullable: true
          type: string
          format: uuid
        cogs_account_id:
          nullable: true
          type: string
          format: uuid
        inventory_account_id:
          nullable: true
          type: string
          format: uuid
        allow_discount:
          type: boolean
        is_active:
          type: boolean
        metadata:
          type: object
...
    PosProfileList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/PosProfile'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          example: 100
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total_pages:
          type: integer
          example: 5
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Invalid input data
        details:
          type: object
          additionalProperties: true
  parameters:
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPageParam:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortByParam:
      name: sort_by
      in: query
      schema:
        type: string
    SortOrderParam:
      name: sort_order
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: asc
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []