# src/presentation/http/pos_profile_handler.rs
//! PosProfile REST handlers
//!
//! Generated by metaphor-schema. Do not edit manually.
//!
//! Uses Axum and backbone-core's BackboneCrudHandler for all 12 CRUD endpoints.

use std::collections::HashMap;
use std::sync::Arc;

use axum::Router;
use serde::{Deserialize, Serialize};
use uuid::Uuid;
use rust_decimal::Decimal;

// Backbone framework imports
use backbone_core::http::BackboneCrudHandler;

// Auth integration (optional)
#[cfg(feature = "auth")]
use backbone_auth::middleware::AuthContext;
#[cfg(feature = "auth")]
use backbone_auth::AuthMiddleware;

// Domain imports
use crate::domain::entity::*;
use crate::application::service::{PosProfileService, ServiceError};

// DTO imports
use crate::presentation::dto::{CreatePosProfileDto, UpdatePosProfileDto, PatchPosProfileDto, PosProfileResponseDto};


/// Application error type
#[derive(Debug, thiserror::Error)]
pub enum PosProfileError {
    #[error("Not found: {0}")]
    NotFound(String),
    #[error("Validation error: {0}")]
    Validation(String),
    #[error("Database error: {0}")]
    Database(String),
    #[error("Internal error: {0}")]
    Internal(String),
}

impl From<ServiceError> for PosProfileError {
    fn from(err: ServiceError) -> Self {
        match err {
            ServiceError::NotFound => Self::NotFound(err.to_string()),
            ServiceError::Validation(ref msg) => Self::Validation(msg.clone()),
            ServiceError::AlreadyExists(ref msg) => Self::Validation(msg.clone()),
            ServiceError::Repository(ref e) => Self::Database(e.to_string()),
            ServiceError::Internal(ref msg) => Self::Internal(msg.clone()),
        }
    }
}

impl axum::response::IntoResponse for PosProfileError {
    fn into_response(self) -> axum::response::Response {
        use axum::http::StatusCode;
        use axum::Json;

        let (status, code) = match &self {
            Self::NotFound(_) => (StatusCode::NOT_FOUND, "POSPROFILE_NOT_FOUND"),
            Self::Validation(_) => (StatusCode::BAD_REQUEST, "POSPROFILE_VALIDATION_ERROR"),
            Self::Database(_) => (StatusCode::INTERNAL_SERVER_ERROR, "POSPROFILE_DATABASE_ERROR"),
            Self::Internal(_) => (StatusCode::INTERNAL_SERVER_ERROR, "POSPROFILE_INTERNAL_ERROR"),
        };

        let body = serde_json::json!({
            "success": false,
            "error": code,
            "message": self.to_string(),
        });

        (status, Json(body)).into_response()
    }
}

// =============================================================================
// Route Configuration
// =============================================================================

/// Create Axum router with all 16 Backbone endpoints for PosProfile.
///
/// # Routes
///
/// | Method | Path | Description |
/// |--------|------|-------------|
/// | GET | /pos_profiles | List with pagination |
/// | POST | /pos_profiles | Create new |
/// | GET | /pos_profiles/:id | Get by ID |
/// | PUT | /pos_profiles/:id | Full update |
/// | PATCH | /pos_profiles/:id | Partial update |
/// | DELETE | /pos_profiles/:id | Soft delete |
/// | POST | /pos_profiles/bulk | Bulk create |
/// | POST | /pos_profiles/upsert | Upsert |
/// | GET | /pos_profiles/trash | List deleted |
/// | POST | /pos_profiles/:id/restore | Restore |
/// | DELETE | /pos_profiles/empty | Empty trash |
/// | GET | /pos_profiles/:id/deleted | Get deleted by ID |
/// | DELETE | /pos_profiles/trash/:id | Permanent delete from trash |
/// | GET | /pos_profiles/count | Count active entities |
/// | GET | /pos_profiles/trash/count | Count deleted entities |
///
/// # Example
///
/// ```text
/// let service = Arc::new(PosProfileService::with_repository(repository));
/// let router = create_pos_profile_routes(service);
/// ```
pub fn create_pos_profile_routes(service: Arc<PosProfileService>) -> Router {
    BackboneCrudHandler::<PosProfileService, PosProfile, CreatePosProfileDto, UpdatePosProfileDto, PosProfileResponseDto>::routes(
        service,
        "/pos_profiles",
    )
}

/// Create Axum router with only the read (GET) endpoints for PosProfile.
///
/// Safe for public, unauthenticated exposure (e.g., reference data).
/// Mutations must be served separately via `create_pos_profile_write_routes`,
/// typically wrapped in an auth middleware layer.
pub fn create_pos_profile_read_routes(service: Arc<PosProfileService>) -> Router {
    BackboneCrudHandler::<PosProfileService, PosProfile, CreatePosProfileDto, UpdatePosProfileDto, PosProfileResponseDto>::read_routes(
        service,
        "/pos_profiles",
    )
}

/// Create Axum router with only the write (mutation) endpoints for PosProfile.
///
/// These routes must NOT be publicly exposed. Wrap them with an auth
/// middleware before nesting into the application router.
pub fn create_pos_profile_write_routes(service: Arc<PosProfileService>) -> Router {
    BackboneCrudHandler::<PosProfileService, PosProfile, CreatePosProfileDto, UpdatePosProfileDto, PosProfileResponseDto>::write_routes(
        service,
        "/pos_profiles",
    )
}

/// Create authenticated routes with auth middleware.
///
/// Requires the `auth` feature flag. The `AuthMiddleware` implementation
/// is responsible for extracting and validating tokens, then providing
/// an `AuthContext` via request extensions.
#[cfg(feature = "auth")]
pub fn create_protected_pos_profile_routes<A: AuthMiddleware + Send + Sync + 'static>(
    service: Arc<PosProfileService>,
    auth: Arc<A>,
) -> Router {
    use axum::middleware;
    use axum::response::IntoResponse;

    let auth_layer = auth.clone();
    create_pos_profile_routes(service)
        .layer(middleware::from_fn(move |mut req: axum::extract::Request, next: axum::middleware::Next| {
            let auth = auth_layer.clone();
            async move {
                let token = req.headers()
                    .get(axum::http::header::AUTHORIZATION)
                    .and_then(|h| h.to_str().ok())
                    .and_then(|raw| raw.strip_prefix("Bearer ").or_else(|| raw.strip_prefix("bearer ")))
                    .unwrap_or("");
                match auth.authenticate(token).await {
                    Ok(ctx) => {
                        req.extensions_mut().insert(ctx);
                        next.run(req).await
                    }
                    Err(_) => {
                        (axum::http::StatusCode::UNAUTHORIZED,
                         axum::Json(serde_json::json!({
                             "success": false,
                             "error": "unauthorized",
                             "message": "Authentication required"
                         }))
                        ).into_response()
                    }
                }
            }
        }))
}